Added y-axis autoscaling feature back
- This feature is key when using the app as it shows imediately the
metric deviations.
- Removed extra y-axis that was showing package counts. The autoscale in
the main y-axis was lost when we added this extra axis. After trying
several different approaches in bokeh to keep the extra axis and have
the autoscaling working on the main axis I have decided to remove
the extra axis.
- Added filter select widget. Another change that messed up with
autoscaling was the decision to show all filters in single plot.
- Implemented workaround to handle datasets with multiple filters until
we dont have a proper dataset model in SQuaSH.